services/chatbot-service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<3.0
PyJWT>=2.9,<3.0
requests>=2.32,<3.0
openai>=1.30,<2.0
google-genai>=1.0
